backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
anthropic==0.43.0
google-genai==1.5.0
pydantic==2.10.4
python-dotenv==1.0.1
playwright==1.49.1
